refactor: 공통 BaseErrorCode 정의 (#153)
