+++ api/api-member/src/main/java/com/pgms/apimember/exception/AdminException.java
@@ -1,11 +1,13 @@
 package com.pgms.apimember.exception;
 
+import com.pgms.coredomain.domain.common.MemberErrorCode;
+
 import lombok.Getter;
 
 @Getter
 public class AdminException extends CustomException {
 
-	public AdminException(CustomErrorCode errorCode) {
+	public AdminException(MemberErrorCode errorCode) {
 		super(errorCode);
 	}
 }
