+++ api/api-member/src/main/java/com/pgms/apimember/exception/CustomException.java
@@ -1,13 +1,15 @@
 package com.pgms.apimember.exception;
 
+import com.pgms.coredomain.domain.common.MemberErrorCode;
+
 import lombok.Getter;
 
 @Getter
 public class CustomException extends RuntimeException {
 
-	protected final CustomErrorCode errorCode;
+	protected final MemberErrorCode errorCode;
 
-	public CustomException(CustomErrorCode errorCode) {
+	public CustomException(MemberErrorCode errorCode) {
 		super(errorCode.getMessage());
 		this.errorCode = errorCode;
 	}
