+++ api/api-member/src/main/java/com/pgms/apimember/exception/GlobalExceptionHandler.java
@@ -1,7 +1,6 @@
 package com.pgms.apimember.exception;
 
 import java.util.List;
-import java.util.Objects;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -11,6 +10,7 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
+import com.pgms.coredomain.domain.common.MemberErrorCode;
 import com.pgms.coredomain.response.ErrorResponse;
 
 import lombok.extern.slf4j.Slf4j;
@@ -29,20 +29,17 @@ public class GlobalExceptionHandler {
 	@ExceptionHandler(CustomException.class)
 	protected ResponseEntity<ErrorResponse> handleEventCustomException(CustomException ex) {
 		log.warn(">>>>> Custom Exception : {}", ex);
-		CustomErrorCode errorCode = ex.getErrorCode();
-		ErrorResponse errorResponse = new ErrorResponse(errorCode.getErrorCode(), errorCode.getMessage());
-		return ResponseEntity.status(errorCode.getStatus()).body(errorResponse);
+		MemberErrorCode errorCode = ex.getErrorCode();
+		return ResponseEntity.status(errorCode.getStatus()).body(errorCode.getErrorResponse());
 	}
 
 	@ExceptionHandler(MethodArgumentNotValidException.class)
 	protected ResponseEntity<ErrorResponse> handleMethodArgumentNotValidException(MethodArgumentNotValidException ex) {
 		log.warn(">>>>> validation Failed : {}", ex);
 		BindingResult bindingResult = ex.getBindingResult();
-		String errorMessage = Objects.requireNonNull(bindingResult.getFieldError())
-			.getDefaultMessage();
 
 		List<FieldError> fieldErrors = bindingResult.getFieldErrors();
-		ErrorResponse errorResponse = new ErrorResponse(CustomErrorCode.VALIDATION_FAILED.getErrorCode(), errorMessage);
+		ErrorResponse errorResponse = MemberErrorCode.VALIDATION_FAILED.getErrorResponse();
 		fieldErrors.forEach(error -> errorResponse.addValidation(error.getField(), error.getDefaultMessage()));
 		return ResponseEntity.status(ex.getStatusCode()).body(errorResponse);
 	}
