+++ api/api-member/src/main/java/com/pgms/apimember/exception/MemberException.java
@@ -1,11 +1,13 @@
 package com.pgms.apimember.exception;
 
+import com.pgms.coredomain.domain.common.MemberErrorCode;
+
 import lombok.Getter;
 
 @Getter
 public class MemberException extends CustomException {
 
-	public MemberException(CustomErrorCode errorCode) {
+	public MemberException(MemberErrorCode errorCode) {
 		super(errorCode);
 	}
 }
