+++ core/core-domain/src/main/java/com/pgms/coredomain/domain/common/BaseErrorCode.java
@@ -1,0 +1,7 @@
+package com.pgms.coredomain.domain.common;
+
+import com.pgms.coredomain.response.ErrorResponse;
+
+public interface BaseErrorCode {
+	ErrorResponse getErrorResponse();
+}
